# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "job-bot",
    "version": "dev"
  },
  "paths": {
...
    "/api/searches": {
      "get": {
        "tags": [
          "collect"
        ],
        "summary": "List Searches",
        "operationId": "listSearches",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Source"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "parse_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ParseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parse Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/searches/{search_id}": {
      "get": {
        "tags": [
          "collect"
        ],
        "summary": "Get Search",
        "operationId": "getSearch",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Search Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/offers": {
      "get": {
        "tags": [
          "collect"
        ],
        "summary": "List Offers",
        "description": "`recent` : dernières offres apparues ; `popular` : offres vues dans le plus d'alertes ;\n`score` : meilleure note de l'IA d'abord (offres non notées à la fin).\n`view` : à examiner, écartées par le filtre, ou toutes. Les autres paramètres sont des\nfiltres d'affichage ; les compteurs par statut et les facettes en tiennent compte.",
        "operationId": "listOffers",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "recent",
                "popular",
                "score"
              ],
              "type": "string",
              "default": "recent",
              "title": "Sort"
            }
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "to_review",
                "filtered_out",
                "later",
                "in_progress",
                "expired",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "View"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Source"
              },
              "default": [],
              "title": "Sources"
            }
          },
          {
            "name": "cantons",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "maxItems": 26,
              "default": [],
              "title": "Cantons"
            }
          },
          {
            "name": "min_rate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rate"
            }
          },
          {
            "name": "external_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "External Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/offers/{offer_id}": {
      "get": {
        "tags": [
          "collect"
        ],
        "summary": "Get Offer",
        "operationId": "getOffer",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collect": {
      "post": {
        "tags": [
          "collect"
        ],
        "summary": "Start Collect",
        "operationId": "startCollect",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectResponse"
                }
              }
            }
          },
          "409": {
            "description": "Collecte non configurée"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectResponse": {
        "properties": {
          "result": {
            "type": "string",
            "enum": [
              "queued",
              "already_queued"
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "CollectResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OfferCounts": {
        "properties": {
          "to_review": {
            "type": "integer",
            "title": "To Review"
          },
          "filtered_out": {
            "type": "integer",
            "title": "Filtered Out"
          },
          "later": {
            "type": "integer",
            "title": "Later"
          },
          "in_progress": {
            "type": "integer",
            "title": "In Progress"
          },
          "expired": {
            "type": "integer",
            "title": "Expired"
          },
          "all": {
            "type": "integer",
            "title": "All"
          }
        },
        "type": "object",
        "required": [
          "to_review",
          "filtered_out",
          "later",
          "in_progress",
          "expired",
          "all"
        ],
        "title": "OfferCounts"
      },
...
      "OfferFacets": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Facet"
            },
            "type": "array",
            "title": "Sources"
          },
          "cantons": {
            "items": {
              "$ref": "#/components/schemas/Facet"
            },
            "type": "array",
            "title": "Cantons"
          }
        },
        "type": "object",
        "required": [
          "sources",
          "cantons"
        ],
        "title": "OfferFacets",
        "description": "Valeurs disponibles pour les filtres, avec le nombre d'offres de chacune."
      },
...
      "OfferPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OfferOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "counts": {
            "$ref": "#/components/schemas/OfferCounts"
          },
          "facets": {
            "$ref": "#/components/schemas/OfferFacets"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "counts",
          "facets"
        ],
        "title": "OfferPage"
      },
...
      "ParseStatus": {
        "type": "string",
        "enum": [
          "parsed",
          "empty",
          "unrecognized",
          "failed"
        ],
        "title": "ParseStatus"
      },
...
      "SearchDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "alert_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Label"
          },
          "parse_status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "results_count": {
            "type": "integer",
            "title": "Results Count"
          },
          "new_offers_count": {
            "type": "integer",
            "title": "New Offers Count"
          },
          "collected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Collected At"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/SearchOffer"
            },
            "type": "array",
            "title": "Offers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "received_at",
          "subject",
          "alert_label",
          "parse_status",
          "parser_version",
          "error",
          "results_count",
          "new_offers_count",
          "collected_at",
          "offers"
        ],
        "title": "SearchDetail"
      },
...
      "SearchOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "alert_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Label"
          },
          "parse_status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "results_count": {
            "type": "integer",
            "title": "Results Count"
          },
          "new_offers_count": {
            "type": "integer",
            "title": "New Offers Count"
          },
          "collected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Collected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "received_at",
          "subject",
          "alert_label",
          "parse_status",
          "parser_version",
          "error",
          "results_count",
          "new_offers_count",
          "collected_at"
        ],
        "title": "SearchOut"
      },
      "SearchPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SearchPage"
      },
...
      "Source": {
        "type": "string",
        "enum": [
          "jobup",
          "indeed",
          "jobroom",
          "unknown"
        ],
        "title": "Source"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}